fix(app): resolve session navigation by target scope
Route session-id navigation through a shared resolver so subsession parent links use the parent session's actual scope instead of the current URL scope. This keeps the prompt dock and subagent footer consistent with cross-scope session navigation.
